+++ matrix/Matrix_SymPacked.hxx
@@ -54,7 +54,7 @@ namespace Seldon
   public:
     // Constructor.
     Matrix_SymPacked();
-    explicit Matrix_SymPacked(int i, int j = 0);
+    explicit Matrix_SymPacked(size_t i, size_t j = 0);
     Matrix_SymPacked(const Matrix_SymPacked<T, Prop, Storage, Allocator>& A);
 
     // Destructor.
@@ -62,26 +62,26 @@ namespace Seldon
     void Clear();
 
     // Basic methods.
-    int GetDataSize() const;
+    size_t GetDataSize() const;
     int64_t GetMemorySize() const;
 
     // Memory management.
     void Reallocate(size_t i, size_t j);
-    void SetData(int i, int j, pointer data);
+    void SetData(size_t i, size_t j, pointer data);
     void Nullify();
 
     // Element access and affectation.
-    reference operator() (int i, int j);
-    const_reference operator() (int i, int j) const;
-    reference Val(int i, int j);
-    const_reference Val(int i, int j) const;
-    reference Get(int i, int j);
-    const_reference Get(int i, int j) const;
-    reference operator[] (int i);
-    const_reference operator[] (int i) const;
+    reference operator() (size_t i, size_t j);
+    const_reference operator() (size_t i, size_t j) const;
+    reference Val(size_t i, size_t j);
+    const_reference Val(size_t i, size_t j) const;
+    reference Get(size_t i, size_t j);
+    const_reference Get(size_t i, size_t j) const;
+    reference operator[] (size_t i);
+    const_reference operator[] (size_t i) const;
     Matrix_SymPacked<T, Prop, Storage, Allocator>&
     operator= (const Matrix_SymPacked<T, Prop, Storage, Allocator>& A);
-    void Set(int i, int j, const T& x);
+    void Set(size_t i, size_t j, const T& x);
     void Copy(const Matrix_SymPacked<T, Prop, Storage, Allocator>& A);
 
     // Convenient functions.
@@ -94,8 +94,8 @@ namespace Seldon
     Matrix_SymPacked<T, Prop, Storage, Allocator>& operator= (const T0& x);
     void FillRand();
     void Print() const;
-    void Print(int a, int b, int m, int n) const;
-    void Print(int l) const;
+    void Print(size_t a, size_t b, size_t m, size_t n) const;
+    void Print(size_t l) const;
 
     // Input/output functions.
     void Write(string FileName) const;
@@ -118,7 +118,7 @@ namespace Seldon
     virtual void ApplySor(const class_SeldonTrans&, Vector<T>& x, const Vector<T>& r,
 			  const typename ClassComplexType<T>::Treal& omega,
 			  int nb_iter, int stage_ssor) const;
-        
+
     virtual void MltAddVector(const Treal& alpha, const Vector<Treal>& x,
 			      const Treal& beta, Vector<Treal>& y) const;
 
@@ -132,16 +132,16 @@ namespace Seldon
     virtual void MltAddVector(const Tcplx& alpha, const SeldonTranspose&,
 			      const Vector<Tcplx>& x,
 			      const Tcplx& beta, Vector<Tcplx>& y) const;
-    
+
     virtual void MltVector(const Vector<Treal>& x, Vector<Treal>& y) const;
     virtual void MltVector(const Vector<Tcplx>& x, Vector<Tcplx>& y) const;
-    
+
     virtual void MltVector(const SeldonTranspose&,
 			   const Vector<Treal>& x, Vector<Treal>& y) const;
 
     virtual void MltVector(const SeldonTranspose&,
 			   const Vector<Tcplx>& x, Vector<Tcplx>& y) const;
-    
+
     virtual bool IsSymmetric() const;
 #endif
 
@@ -162,7 +162,7 @@ namespace Seldon
 
   public:
     Matrix();
-    explicit Matrix(int i, int j = 0);
+    explicit Matrix(size_t i, size_t j = 0);
 
     template <class T0>
     Matrix<T, Prop, ColSymPacked, Allocator>& operator= (const T0& x);
@@ -171,7 +171,7 @@ namespace Seldon
                                                          Allocator>& A);
     template<class T0>
     Matrix<T, Prop, ColSymPacked, Allocator>& operator*= (const T0& x);
-    void Resize(int i, int j);
+    void Resize(size_t i, size_t j);
   };
 
 
@@ -189,7 +189,7 @@ namespace Seldon
 
   public:
     Matrix();
-    Matrix(int i, int j = 0);
+    Matrix(size_t i, size_t j = 0);
 
     template <class T0>
     Matrix<T, Prop, RowSymPacked, Allocator>& operator= (const T0& x);
@@ -198,7 +198,7 @@ namespace Seldon
                                                          Allocator>& A);
     template<class T0>
     Matrix<T, Prop, RowSymPacked, Allocator>& operator*= (const T0& x);
-    void Resize(int i, int j);
+    void Resize(size_t i, size_t j);
   };
